requirements.txt
pillow>=10.0.0
requests>=2.31.0
schedule>=1.2.0
selenium>=4.15.0
